move MacronutrientsTable into its own file
